set filtering and pagination from data
